폴더 구조 최적화 작업 및 홈페이지 기능 추가 및 수정: - a 링크 새 탭 이동 - cover 섹션 이미지 클릭 시 이동하는 링크 주소 수정 - 요일별 연재 부분 수정
