+++ src/features/home/home.html
@@ -31,10 +31,8 @@
           <!-- 표지(커버) 섹션 시작 -->
           <section class="cover"></section>
 
-          <!--이 섹션은 구현 안 해도 되지만 다른 방식으로 구현 해보고 싶어서..-->
-          <!-- 요일별 연재 -> 요일별 모아보기 섹션 시작 -->
+          <!-- 요일별 연재 섹션 시작 -->
           <section class="weekly-serial" aria-labelledby="weeklySerialTitle">
-            <!-- 요일별 연재 섹션 시작 -->
             <h2 class="weekly-serial__heading" id="weeklySerialTitle">
               요일별 연재
             </h2>
@@ -366,6 +364,7 @@
                     href="https://www.facebook.com/brunch.co.kr/?locale=ko_KR"
                     class="social__link"
                     aria-label="공식 페이스북"
+                    target="_blank"
                   >
                     <img
                       src="/assets/icons/social/social_fb.svg"
@@ -379,6 +378,7 @@
                     href="#"
                     class="social__link"
                     aria-label="공식 x(구 트위터)"
+                    target="_blank"
                   >
                     <img
                       src="/assets/icons/social/social_x.svg"
@@ -400,6 +400,7 @@
                     <a
                       href="https://www.kakao.com/policy/terms"
                       class="main__legal--link"
+                      target="_blank"
                       >이용약관</a
                     >
                   </li>
@@ -407,6 +408,7 @@
                     <a
                       href="https://brunch.co.kr/policy/terms"
                       class="main__legal--link"
+                      target="_blank"
                       >이전 이용약관</a
                     >
                   </li>
@@ -414,6 +416,7 @@
                     <a
                       href="https://billing-web.kakao.com/kbill/terms/service"
                       class="main__legal--link"
+                      target="_blank"
                       >카카오 유료서비스 이용약관</a
                     >
                   </li>
@@ -423,6 +426,7 @@
                     <a
                       href="https://www.kakao.com/policy/privacy?lang=ko"
                       class="main__legal--link"
+                      target="_blank"
                       >카카오 개인정보 처리방침</a
                     >
                   </li>
@@ -430,6 +434,7 @@
                     <a
                       href="https://www.kakao.com/policy/safeguard?lang=ko"
                       class="main__legal--link"
+                      target="_blank"
                       >청소년 보호정책</a
                     >
                   </li>
@@ -437,6 +442,7 @@
                     <a
                       href="https://brunch.co.kr/policy/oppolicy"
                       class="main__legal--link"
+                      target="_blank"
                       >운영정책</a
                     >
                   </li>
